requirements.txt
asgiref==3.7.2
Django==5.0.1
django-braces==1.15.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
